requirements.txt
langchain-openai==0.3.9
langchain==0.3.27
pydantic==2.11.5
playwright==1.55.0
mcp==1.13.1
beautifulsoup4==4.13.4
readability-lxml==0.8.4.1
jinjia2==3.1.6
packaging==24.2
